+++ CHANGELOG.md
@@ -1,2 +1,10 @@
+## Release 0.0.2
+  **NEW CONTENT**
+- Animations for ritual spells based on spell school.
+- Animations for 'Other' and 'Utility' spells based on conditions, or disposition colored by spell school.
+  **BUGFIXES**
+- Save animation for spells without damage.
+- Save animation outro file.
+  
 ## Alpha Release (v0.0.1)
 - All initial content, see [ReadMe](README.md)